+++ test/e2e/non-ascii-headers/without-middleware/app/page.js
@@ -1,0 +1,3 @@
+export default function Page() {
+  return <div>Hello</div>
+}
