+++ test/e2e/non-ascii-headers/without-middleware/next.config.js
@@ -1,0 +1,1 @@
+module.exports = {}
